Factor out CodeMonitorForm from CreateCodeMonitorPage
